Refactor/files

+++ backend/src/main/java/com/bookhub/bookhub_back/dto/statistics/response/stocks/TimeStockChartResponseDto.java
@@ -11,7 +11,7 @@ import lombok.Setter;
 @Setter
 public class TimeStockChartResponseDto {
     private String branchName;
-    private Long month;
+    private int month;
     private Long inAmount;
     private Long lossAmount;
 }
